Classes to help hackerboatStateClassStorable instances read and write themselves to/from an SQLite database.
In particular, hackerboatStateStorage represents one storage location: a
database connection handle, a table name, enough information about the
schema for reading and writing, and a collected of cached prepared
SQLite queries. The expectation is that each concrete subclass of
hackerboatStateClassStorable will correspond to one instance of
hackerboatStateStorage managing that class's stored values.